Defer decision until workflow-start time
